Remove harmonically bounded candidates
Remove harmonically bounded candidates when calculating grammars. Those
candidates are stored separately, and can be retrieved with the
get_harmonically_bounded_candidates method. The API has changed for
that, as instead of pairs of candidates it now returns just the bounded
candidate.

Also refactor the get_entails method to make it more readable.