Add a-priori filtering to PoOT
Specify a frozenset ranking to the poot.PoOT constructor to filter
grammars and entailments based on that ranking. Be sure that the ranking
given is a frozenset of tuples of the form (a, b) where a IS DOMINATED
BY b. Otherwise you will be asking to filter by the inverse ranking.